refactor: refine statistical approach for tallying posts by category association
