+++ api/src/main/java/run/halo/app/core/extension/content/Post.java
@@ -42,6 +42,8 @@ public class Post extends AbstractExtension {
     public static final String LAST_RELEASED_SNAPSHOT_ANNO =
         "content.halo.run/last-released-snapshot";
     public static final String LAST_ASSOCIATED_TAGS_ANNO = "content.halo.run/last-associated-tags";
+    public static final String LAST_ASSOCIATED_CATEGORIES_ANNO =
+        "content.halo.run/last-associated-categories";
 
     public static final String DELETED_LABEL = "content.halo.run/deleted";
     public static final String PUBLISHED_LABEL = "content.halo.run/published";
